# api/src/main/resources/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "This is API definition for the My Application",
    "version": "1.0.0-SNAPSHOT",
    "title": "My App"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "login-controller",
      "description": "Login Controller"
    },
    {
      "name": "orders-controller",
      "description": "Orders Controller"
    }
  ],
  "paths": {
    "/auth/urls": {
      "get": {
        "tags": [
          "login-controller"
        ],
        "summary": "Fetches all oauth login urls",
        "operationId": "getOauthUrls",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schema/oauthUrl.schema.json"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "login-controller"
        ],
        "summary": "Attempts logging the user in",
        "operationId": "authenticateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schema/requests/loginRequest.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/authResponse.schema.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "login-controller"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Attempts logging the user out",
        "operationId": "logOut",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register-user": {
      "post": {
        "tags": [
          "login-controller"
        ],
        "summary": "Registers a new user manually",
        "operationId": "registerUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schema/requests/signupRequest.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "get": {
        "tags": [
          "login-controller"
        ],
        "summary": "Verifies an email address",
        "operationId": "verifyEmail",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": [
          "orders-controller"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Fetches all orders for the authenticated user",
        "operationId": "fetchOrders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "./schema/order.schema.json"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Places a new order",
        "operationId": "placeOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schema/requests/placeOrderRequest.schema.json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/placeOrderResponse.schema.json"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./schema/responses/errorResponse.schema.json"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "//localhost:8080/api/v1"
    }
  ]
}